version changed to snapshot, geographtools version changed to snapshot, geographtools related changes (edge length and posted speed changed units)

+++ src/main/java/cz/cvut/fel/aic/graphimporter/osm/OsmImporter.java
@@ -406,8 +406,8 @@ public class OsmImporter extends Importer implements OsmElementConsumer {
 		}
 	}
 
-	protected float extractSpeed(OsmWay way, EdgeType edgeType) {
-		return edgeType.apply(speedExtractor, way.getTags()).floatValue();
+	protected int extractSpeed(OsmWay way, EdgeType edgeType) {
+		return edgeType.apply(speedExtractor, way.getTags()).intValue();
 	}
 
 	/**
